# Improve _aca_packets_to_table speed by factor of 20